move options testing into it's own component so the context values are private
